Podcasting: reorganize into components/ and hooks/ subdirectories
